requirements.txt
mysql-connector-python
pandas
firebase_admin